docs: update from latest tilt.build and v0.23.2
This also updates the Tiltfile to eliminate manual steps.

There's also a change to `gendocs` to make the `sed` invocation
work across macOS + Linux since that's the theme of the day.

Something odd is going on with the `Dockerfile` setup and I kept
having bundler/gem issues - needed to also install for the non
base, which seems wrong, but I'm not totally sure what's going on
because it's been so long since I dealt with Ruby/Jekyll.